Add ClusterPerf 1M virtual client benchmark.
Benchmarks "writeThroughputMMC" and "linearizableWriteThroughputMMC"
(Massively Multi Client) are added to clusterperf.  At the core, both
benchmarks are throughput benchmarks.  The primary difference between
MMC and non-MMC version of the benchmarks is that the MMC simulates
the effect of many clients in single CluterPerf client executable.
Each client executable "runs" multiple "Virtual" clients concurrently.
The goal is to allow virtual clients to send an RPC concurrently with
any other virtual clients.

Currently the benchmark runs up to 1M virtual clients targeting a
cluster of 10k Virtual RAMCloud servers a small number of which are
backed by real servers.  Virtual clients randomly selects a server to
target for each RPC.  If the target server is backed by a real server,
an RPC is sent; otherwise, the virtual client simply waits for some
amount of time before trying to send another RPC.  The wait time is
drawn from an empirically collected distribution of RPC latencies so
as to simulate the actual delay of sending an RPC.

In this setup, each real server can expect on average 100 concurrent
RPCs.  At this time, the primary bottleneck is the network stack which
cannot handle this level of concurrency.